workflows: remove time-consuming Konan cache task
In one of the workflows, it took as long as 5m 15s to restore a cache of
Konan [1]. The Dokka deployment task itself doesn't cache Konan and only
took 5m 9s to complete. There's no point to caching Konan, especially
because installing the Kotlin/Native compiler only takes a few seconds
anyway.

[1] https://github.com/inthewaves/kotlin-signald/runs/3502202516#step:4:18